+++ deployment/multiclient.go
@@ -13,6 +13,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/google/uuid"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 
@@ -30,14 +31,18 @@ const (
 )
 
 type RetryConfig struct {
-	Attempts uint
-	Delay    time.Duration
+	Attempts     uint
+	Delay        time.Duration
+	DialAttempts uint
+	DialDelay    time.Duration
 }
 
 func defaultRetryConfig() RetryConfig {
 	return RetryConfig{
-		Attempts: RPCDefaultRetryAttempts,
-		Delay:    RPCDefaultRetryDelay,
+		Attempts:     RPCDefaultRetryAttempts,
+		Delay:        RPCDefaultRetryDelay,
+		DialAttempts: RPCDefaultDialRetryAttempts,
+		DialDelay:    RPCDefaultDialRetryDelay,
 	}
 }
 
@@ -63,6 +68,12 @@ func NewMultiClient(lggr logger.Logger, rpcsCfg RPCConfig, opts ...func(client *
 	}
 	mc := MultiClient{lggr: lggr, chainName: chain.Name}
 
+	mc.RetryConfig = defaultRetryConfig()
+
+	for _, opt := range opts {
+		opt(&mc)
+	}
+
 	clients := make([]*ethclient.Client, 0, len(rpcsCfg.RPCs))
 	for i, rpc := range rpcsCfg.RPCs {
 		client, err := mc.dialWithRetry(rpc, lggr)
@@ -79,11 +90,7 @@ func NewMultiClient(lggr logger.Logger, rpcsCfg RPCConfig, opts ...func(client *
 
 	mc.Client = clients[0]
 	mc.Backups = clients[1:]
-	mc.RetryConfig = defaultRetryConfig()
 
-	for _, opt := range opts {
-		opt(&mc)
-	}
 	return &mc, nil
 }
 
@@ -98,8 +105,10 @@ func (mc *MultiClient) CallContract(ctx context.Context, msg ethereum.CallMsg, b
 	err := mc.retryWithBackups("CallContract", func(client *ethclient.Client) error {
 		var err error
 		result, err = client.CallContract(ctx, msg, blockNumber)
+
 		return err
 	})
+
 	return result, err
 }
 
@@ -108,8 +117,10 @@ func (mc *MultiClient) CallContractAtHash(ctx context.Context, msg ethereum.Call
 	err := mc.retryWithBackups("CallContractAtHash", func(client *ethclient.Client) error {
 		var err error
 		result, err = client.CallContractAtHash(ctx, msg, blockHash)
+
 		return err
 	})
+
 	return result, err
 }
 
@@ -118,8 +129,10 @@ func (mc *MultiClient) CodeAt(ctx context.Context, account common.Address, block
 	err := mc.retryWithBackups("CodeAt", func(client *ethclient.Client) error {
 		var err error
 		code, err = client.CodeAt(ctx, account, blockNumber)
+
 		return err
 	})
+
 	return code, err
 }
 
@@ -128,8 +141,10 @@ func (mc *MultiClient) CodeAtHash(ctx context.Context, account common.Address, b
 	err := mc.retryWithBackups("CodeAtHash", func(client *ethclient.Client) error {
 		var err error
 		code, err = client.CodeAtHash(ctx, account, blockHash)
+
 		return err
 	})
+
 	return code, err
 }
 
@@ -138,8 +153,10 @@ func (mc *MultiClient) NonceAt(ctx context.Context, account common.Address, bloc
 	err := mc.retryWithBackups("NonceAt", func(client *ethclient.Client) error {
 		var err error
 		count, err = client.NonceAt(ctx, account, block)
+
 		return err
 	})
+
 	return count, err
 }
 
@@ -148,8 +165,10 @@ func (mc *MultiClient) NonceAtHash(ctx context.Context, account common.Address, 
 	err := mc.retryWithBackups("NonceAtHash", func(client *ethclient.Client) error {
 		var err error
 		count, err = client.NonceAtHash(ctx, account, blockHash)
+
 		return err
 	})
+
 	return count, err
 }
 
@@ -158,8 +177,10 @@ func (mc *MultiClient) HeaderByNumber(ctx context.Context, number *big.Int) (*ty
 	err := mc.retryWithBackups("HeaderByNumber", func(client *ethclient.Client) error {
 		var err error
 		header, err = client.HeaderByNumber(ctx, number)
+
 		return err
 	})
+
 	return header, err
 }
 
@@ -191,30 +212,35 @@ func (mc *MultiClient) WaitMined(ctx context.Context, tx *types.Transaction) (*t
 	case receipt = <-resultCh:
 		close(doneCh)
 		mc.lggr.Debugf("Tx %s mined with chain %s", tx.Hash().Hex(), mc.chainName)
+
 		return receipt, nil
 	case <-ctx.Done():
 		mc.lggr.Warnf("WaitMined context done %v", ctx.Err())
 		close(doneCh)
+
 		return nil, ctx.Err()
 	}
 }
 
 func (mc *MultiClient) retryWithBackups(opName string, op func(*ethclient.Client) error) error {
 	var err error
+	traceID := uuid.New()
 	for i, client := range append([]*ethclient.Client{mc.Client}, mc.Backups...) {
 		err2 := retry.Do(func() error {
 			err = op(client)
 			if err != nil {
-				mc.lggr.Warnf("retryable error '%s' for op %s with chain %s client index %d", MaybeDataErr(err), opName, mc.chainName, i)
+				mc.lggr.Warnf("traceID(%s): retryable error '%s' for op %s with chain %s client index %d", traceID.String(), MaybeDataErr(err), opName, mc.chainName, i)
 				return err
 			}
+
 			return nil
 		}, retry.Attempts(mc.RetryConfig.Attempts), retry.Delay(mc.RetryConfig.Delay))
 		if err2 == nil {
 			return nil
 		}
-		mc.lggr.Infof("Client at index %d failed, trying next client chain %s", i, mc.chainName)
+		mc.lggr.Infof("traceID(%s): Client at index %d failed, trying next client chain %s", traceID.String(), i, mc.chainName)
 	}
+
 	return errors.Join(err, fmt.Errorf("all backup clients failed for chain %s", mc.chainName))
 }
 
@@ -224,21 +250,24 @@ func (mc *MultiClient) dialWithRetry(rpc RPC, lggr logger.Logger) (*ethclient.Cl
 		return nil, err
 	}
 
+	traceID := uuid.New()
 	var client *ethclient.Client
 	err = retry.Do(func() error {
 		var err2 error
-		mc.lggr.Debugf("dialing endpoint '%s' for RPC %s for chain %s", endpoint, rpc.Name, mc.chainName)
+		mc.lggr.Debugf("traceID(%s): dialing endpoint '%s' for RPC %s for chain %s", traceID.String(), endpoint, rpc.Name, mc.chainName)
 		client, err2 = ethclient.Dial(endpoint)
 		if err2 != nil {
-			lggr.Warnf("retryable error for RPC %s:%s for chain %s  %v", rpc.Name, endpoint, mc.chainName, err2)
+			lggr.Warnf("traceID(%s): retryable error for RPC %s:%s for chain %s  %v", traceID.String(), rpc.Name, endpoint, mc.chainName, err2)
 			return err2
 		}
+
 		return nil
-	}, retry.Attempts(RPCDefaultDialRetryAttempts), retry.Delay(RPCDefaultDialRetryDelay))
+	}, retry.Attempts(mc.RetryConfig.DialAttempts), retry.Delay(mc.RetryConfig.DialDelay))
 
 	if err != nil {
 		return nil, errors.Join(err, fmt.Errorf("failed to dial endpoint '%s' for RPC %s for chain %s after retries", endpoint, rpc.Name, mc.chainName))
 	}
+
 	return client, nil
 }
 
@@ -250,8 +279,10 @@ func (mc *MultiClient) SuggestGasPrice(ctx context.Context) (*big.Int, error) {
 	err := mc.callWithDebug("SuggestGasPrice", func(client *ethclient.Client) error {
 		var err error
 		gasPrice, err = client.SuggestGasPrice(ctx)
+
 		return err
 	})
+
 	return gasPrice, err
 }
 
@@ -260,8 +291,10 @@ func (mc *MultiClient) SuggestGasTipCap(ctx context.Context) (*big.Int, error) {
 	err := mc.callWithDebug("SuggestGasTipCap", func(client *ethclient.Client) error {
 		var err error
 		gasTipCap, err = client.SuggestGasTipCap(ctx)
+
 		return err
 	})
+
 	return gasTipCap, err
 }
 
@@ -270,8 +303,10 @@ func (mc *MultiClient) PendingCodeAt(ctx context.Context, account common.Address
 	err := mc.callWithDebug("PendingCodeAt", func(client *ethclient.Client) error {
 		var err error
 		code, err = client.PendingCodeAt(ctx, account)
+
 		return err
 	})
+
 	return code, err
 }
 
@@ -280,8 +315,10 @@ func (mc *MultiClient) PendingNonceAt(ctx context.Context, account common.Addres
 	err := mc.callWithDebug("PendingNonceAt", func(client *ethclient.Client) error {
 		var err error
 		count, err = client.PendingNonceAt(ctx, account)
+
 		return err
 	})
+
 	return count, err
 }
 
@@ -290,8 +327,10 @@ func (mc *MultiClient) EstimateGas(ctx context.Context, call ethereum.CallMsg) (
 	err := mc.callWithDebug("EstimateGas", func(client *ethclient.Client) error {
 		var err error
 		gas, err = client.EstimateGas(ctx, call)
+
 		return err
 	})
+
 	return gas, err
 }
 
@@ -300,8 +339,10 @@ func (mc *MultiClient) BalanceAt(ctx context.Context, account common.Address, bl
 	err := mc.callWithDebug("BalanceAt", func(client *ethclient.Client) error {
 		var err error
 		balance, err = client.BalanceAt(ctx, account, blockNumber)
+
 		return err
 	})
+
 	return balance, err
 }
 
@@ -310,8 +351,10 @@ func (mc *MultiClient) FilterLogs(ctx context.Context, q ethereum.FilterQuery) (
 	err := mc.callWithDebug("FilterLogs", func(client *ethclient.Client) error {
 		var err error
 		logs, err = client.FilterLogs(ctx, q)
+
 		return err
 	})
+
 	return logs, err
 }
 
@@ -320,17 +363,19 @@ func (mc *MultiClient) SubscribeFilterLogs(ctx context.Context, q ethereum.Filte
 	err := mc.callWithDebug("SubscribeFilterLogs", func(client *ethclient.Client) error {
 		var err error
 		sub, err = client.SubscribeFilterLogs(ctx, q, ch)
+
 		return err
 	})
+
 	return sub, err
 }
 
 func (mc *MultiClient) callWithDebug(callName string, call func(*ethclient.Client) error) error {
 	mc.lggr.Debugf("Calling %s for chain %s", callName, mc.chainName)
-	var err error
-	err = call(mc.Client)
+	var err = call(mc.Client)
 	if err != nil {
 		mc.lggr.Errorf("error '%s' when calling %s for chain %s", MaybeDataErr(err), callName, mc.chainName)
 	}
+
 	return err
 }