+++ deployment/multiclient_test.go
@@ -14,6 +14,7 @@ import (
 
 // TODO(giogam): This test is incomplete, it should be completed with support for websockets URLS
 func TestMultiClient(t *testing.T) {
+	t.Parallel()
 	var (
 		lggr                 = logger.Test(t)
 		chainSelector uint64 = 16015286601757825753 // "ethereum-testnet-sepolia"
@@ -30,8 +31,10 @@ func TestMultiClient(t *testing.T) {
 	require.NotNil(t, mc)
 
 	assert.Equal(t, "ethereum-testnet-sepolia", mc.chainName)
-	assert.Equal(t, mc.RetryConfig.Attempts, uint(RPCDefaultRetryAttempts))
+	assert.Equal(t, uint(RPCDefaultRetryAttempts), mc.RetryConfig.Attempts)
 	assert.Equal(t, RPCDefaultRetryDelay, mc.RetryConfig.Delay)
+	assert.Equal(t, uint(RPCDefaultDialRetryAttempts), mc.RetryConfig.DialAttempts)
+	assert.Equal(t, RPCDefaultDialRetryDelay, mc.RetryConfig.DialDelay)
 
 	// Expect error if no RPCs provided.
 	_, err = NewMultiClient(lggr, RPCConfig{ChainSelector: chainSelector, RPCs: []RPC{}})
@@ -93,10 +96,10 @@ func TestMultiClient_retryWithBackups(t *testing.T) {
 			// Run operation and check expectations
 			err = mc.retryWithBackups(tt.opName, tt.op)
 			if tt.wantErr != "" {
-				assert.Error(t, err)
-				assert.Equal(t, tt.wantErr, err.Error())
+				require.Error(t, err)
+				require.ErrorContains(t, err, tt.wantErr)
 			} else {
-				assert.NoError(t, err)
+				require.NoError(t, err)
 			}
 		})
 	}